+++ __tmp/patch.py
@@ -1,0 +1,2 @@
+import base64, sys
+exec(base64.b64decode(sys.argv[1]).decode())